feat#49: rearranged utils package into separate packages with utilitary usage - jsonx - clone - console - maps - markdown - pathx
