+++ cmd/ada-cli/main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/baudii/ada-ai/internal/cli"
 	"github.com/baudii/ada-ai/internal/common"
 	"github.com/baudii/ada-ai/internal/project/folder"
-	"github.com/baudii/ada-ai/pkg/utils"
+	"github.com/baudii/ada-ai/pkg/must"
 )
 
 func main() {
@@ -26,12 +26,12 @@ func main() {
 	// Set up logging
 	cfgPath := filepath.Join(common.ConfigPath, "dilog.json")
 	cfg := common.LoadLogConfig(cfgPath)
-	logger := utils.Must(common.DefaultSimpleLogger(cfg))
+	logger := must.Value(common.DefaultSimpleLogger(&cfg))
 	slog.SetDefault(logger)
 	logger.Info("initialized logger", "config", cfgPath)
 
 	// Create and run the application
-	opts := utils.Must(app.ParseAppOptions(common.ConfigPath))
+	opts := must.Value(app.ParseAppOptions(common.ConfigPath))
 	path := filepath.Join(common.AiConfigPath, fmt.Sprintf("%s.json", *provider))
 	c := cli.New().GetProjectData()
 	app := app.New(
@@ -41,8 +41,8 @@ func main() {
 		app.WithProjectData(c),
 		app.WithDirProvider(folder.NewSeqDir(os.ReadDir)),
 	)
-	utils.MustErr(app.InitAda(*provider, path))
-	utils.MustErr(app.InitLocalProject())
+	must.Do(app.InitAda(*provider, path))
+	must.Do(app.InitLocalProject())
 	if err := app.Run(context.Background()); err != nil {
 		log.Fatalf("runtime error: %v", err)
 	}
