+++ internal/cli/cli.go
@@ -6,7 +6,8 @@ import (
 
 	"github.com/baudii/ada-ai/internal/app"
 	"github.com/baudii/ada-ai/internal/common"
-	"github.com/baudii/ada-ai/pkg/utils"
+	"github.com/baudii/ada-ai/pkg/console"
+	"github.com/baudii/ada-ai/pkg/jsonx"
 )
 
 type cliApp struct {
@@ -15,7 +16,7 @@ type cliApp struct {
 
 // New creates a new instance of the CLI application that implements the Runner interface.
 func New() *cliApp {
-	return &cliApp{utils.ReadInput}
+	return &cliApp{console.ReadInput}
 }
 
 // GetProjectData retrieves project data from a JSON file or prompts the user for input
@@ -23,13 +24,13 @@ func New() *cliApp {
 // through the provided channel and closes the channel when done.
 func (cli *cliApp) GetProjectData() app.ProjectData {
 	path := filepath.Join(common.DataPath, "user_data.json")
-	projectData, err := utils.ParseJSONFile[app.ProjectData](path)
+	projectData, err := jsonx.Load[app.ProjectData](path)
 	if err != nil {
 		username := cli.read("Provide nickname")
 		projname := cli.read("Provide project name")
 		plang := cli.read("Provide programming language (go, python, js, etc)")
 		summary := cli.read("Provide a short summary of the project")
-		projectData = &app.ProjectData{
+		projectData = app.ProjectData{
 			UserName: username,
 			ProjName: projname,
 			Language: plang,
@@ -38,10 +39,10 @@ func (cli *cliApp) GetProjectData() app.ProjectData {
 
 	}
 
-	err = utils.SaveJSONToFile(projectData, path)
+	err = jsonx.Save(projectData, path)
 	if err != nil {
 		slog.Error("failed to save project data to file", "path", path, "error", err)
 	}
 
-	return *projectData
+	return projectData
 }
