+++ pkg/must/must_test.go
@@ -1,4 +1,4 @@
-package utils
+package must
 
 import (
 	"errors"
@@ -21,10 +21,10 @@ func TestMust(t *testing.T) {
 	for _, v := range tests {
 		t.Run(fmt.Sprintf("%v", v.val), func(t *testing.T) {
 			if v.err == nil {
-				result := Must(v.val, v.err)
+				result := Value(v.val, v.err)
 				assert.Equal(t, v.val, result)
 			} else {
-				assert.PanicsWithValue(t, v.err.Error(), func() { Must(v.val, errors.New("err")) })
+				assert.PanicsWithValue(t, v.err.Error(), func() { Value(v.val, errors.New("err")) })
 			}
 		})
 	}
@@ -41,9 +41,9 @@ func TestMustErr(t *testing.T) {
 	for _, v := range tests {
 		t.Run(fmt.Sprintf("%v", v.err), func(t *testing.T) {
 			if v.err == nil {
-				MustErr(nil)
+				Do(nil)
 			} else {
-				assert.PanicsWithValue(t, v.err.Error(), func() { MustErr(errors.New("err")) })
+				assert.PanicsWithValue(t, v.err.Error(), func() { Do(errors.New("err")) })
 			}
 		})
 	}
