--- pkg/utils/utils_test.go
@@ -1,236 +1,0 @@
-package utils
-
-import (
-	"bytes"
-	"fmt"
-	"io"
-	"path/filepath"
-	"strconv"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-type mockWriter struct {
-	f func() (int, error)
-}
-
-func (m *mockWriter) Write(p []byte) (n int, err error) {
-	return m.f()
-}
-
-func TestAbsolutePath(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		relPath  string
-		basePath string
-		err      error
-	}{
-		{"anything", t.TempDir(), nil},
-		{"a/b/c/", t.TempDir(), nil},
-		{"/a/s/d", t.TempDir(), nil},
-		{"", t.TempDir(), nil},
-		{"", "", fmt.Errorf("failed to get base")},
-	}
-
-	for i, v := range tests {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			v.basePath = filepath.Join(v.basePath, "exe.exe")
-			executable := func() (string, error) { return v.basePath, v.err }
-			if v.err != nil {
-				assert.PanicsWithError(t, v.err.Error(), func() { AbsolutePath(v.relPath, executable) })
-			} else {
-				ap := AbsolutePath(v.relPath, executable)
-				expected := filepath.Join(filepath.Dir(v.basePath), v.relPath)
-				assert.Equal(t, expected, ap)
-			}
-		})
-	}
-}
-
-func TestDeepCopyMap(t *testing.T) {
-	t.Parallel()
-	s1 := struct{ s string }{"hello"}
-	s2 := struct{ s any }{"world"}
-	var (
-		nilSl  []int          = nil
-		nilPtr *int           = nil
-		nilMap map[string]any = nil
-		nilIfc any
-	)
-	m1 := map[string]any{
-		"a": "bbb",
-		"b": map[string]any{"c": 123, "s": nil},
-		"d": []int{1, 2, 3},
-		"e": nil,
-		"f": &s1,
-		"g": nilSl,
-		"h": nilMap,
-		"i": nilPtr,
-		"j": nilIfc,
-	}
-
-	m3 := DeepCopyMap(nil)
-	m2 := DeepCopyMap(m1)
-	m1["a"] = "zzz"
-	m1["d"].([]int)[0] = 100
-	delete(m1, "b")
-	m1["e"] = "hello"
-	m1["f"] = &s2
-
-	assert.Nil(t, m3)
-	assert.Equal(t, map[string]any{"a": "zzz", "d": []int{100, 2, 3}, "e": "hello", "f": &s2, "g": nilSl, "h": nilMap, "i": nilPtr, "j": nilIfc}, m1)
-	assert.Equal(t, map[string]any{
-		"a": "bbb",
-		"b": map[string]any{"c": 123, "s": nil},
-		"d": []int{1, 2, 3},
-		"e": nil,
-		"f": &s1,
-		"g": nilSl,
-		"h": nilMap,
-		"i": nilPtr,
-		"j": nilIfc,
-	}, m2)
-}
-
-func TestMergeMap(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		dst      map[string]any
-		src      map[string]any
-		expected map[string]any
-	}{
-		{nil, map[string]any{"a": nil}, nil},
-		{map[string]any{"b": "s"}, nil, map[string]any{"b": "s"}},
-		{map[string]any{"a": "s"}, map[string]any{"a": nil}, map[string]any{"a": nil}},
-		{map[string]any{"a": "s"}, map[string]any{"b": nil}, map[string]any{"a": "s", "b": nil}},
-		{
-			map[string]any{
-				"nil1": nil,
-				"int":  12,
-				"e": map[string]any{
-					"nil2":     nil,
-					"ovwrtSt":  "st",
-					"ovwrtInt": 12,
-				},
-			},
-			map[string]any{"e": map[string]any{
-				"nil2":     "not nil",
-				"ovwrtSt":  12,
-				"ovwrtInt": "st",
-				"new":      "new",
-			}},
-			map[string]any{
-				"nil1": nil,
-				"int":  12,
-				"e": map[string]any{
-					"nil2":     "not nil",
-					"ovwrtSt":  12,
-					"ovwrtInt": "st",
-					"new":      "new",
-				},
-			},
-		},
-	}
-
-	for i, v := range tests {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			MergeMap(v.dst, v.src)
-			assert.Equal(t, v.dst, v.expected)
-		})
-	}
-}
-
-func TestPrintTree(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		m        map[string]any
-		expected []string
-		hasErr   bool
-	}{
-		{map[string]any{"a": map[string]any{"b": 0}}, []string{"└── a\n    └── b\n"}, false},
-		{map[string]any{"a": map[string]any{"b": 0, "c": 0}, "b": map[string]any{"c": 0}}, []string{
-			"├── a\n│   ├── c\n│   └── b\n└── b\n    └── c\n",
-			"├── a\n│   ├── b\n│   └── c\n└── b\n    └── c\n",
-			"├── b\n│   └── c\n└── a\n    ├── b\n    └── c\n",
-			"├── b\n│   └── c\n└── a\n    ├── c\n    └── b\n",
-		}, false},
-		{map[string]any{"a": map[string]any{"b": 0, "c": 0}, "b": map[string]any{"c": 0}}, nil, true},
-	}
-
-	for i, v := range tests {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			iter := 0
-			var w io.Writer = &bytes.Buffer{}
-			if v.hasErr {
-				w = &mockWriter{f: func() (int, error) {
-					if iter > 0 {
-						return 0, fmt.Errorf("some err")
-					}
-					iter++
-					return 0, nil
-				}}
-			}
-			err := PrintTree(w, v.m, "")
-			if v.hasErr {
-				assert.Error(t, err)
-				return
-			}
-			b, ok := w.(*bytes.Buffer)
-			if !assert.True(t, ok) {
-				return
-			}
-			assert.Contains(t, v.expected, b.String())
-		})
-	}
-}
-
-func TestUlistMd(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		input    []any
-		expected string
-	}{
-		{[]any{"one"}, "* one"},
-		{[]any{"one", 1, true}, "* one\n* 1\n* true"},
-		{[]any{}, ""},
-		{nil, ""},
-	}
-	for i, v := range tests {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			assert.Equal(t, v.expected, UlistMd(v.input...))
-		})
-	}
-}
-
-func TestStringify(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		input    any
-		expected string
-	}{
-		{struct{ A string }{"foo"}, "- A: foo"},
-		{struct {
-			A string
-			B int
-		}{"foo", 42}, "- A: foo\n- B: 42"},
-		{struct {
-			A *any
-			B *any
-		}{nil, nil}, "- A: <nil>\n- B: <nil>"},
-		{struct {
-			A string
-			B int
-		}{"foo", 0}, "- A: foo\n- B: 0"},
-		{struct {
-			A string
-			B int
-		}{"", 42}, "- A: \n- B: 42"},
-	}
-
-	for i, v := range tests {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			assert.Equal(t, v.expected, Stringify(v.input))
-		})
-	}
-}
